pyproject.toml
[tool.poetry]
name = "youtubemaster"
version = "0.1.0"
description = "A GUI frontend for yt-dlp"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "youtubemaster", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<=3.13.2"
pyinstaller = "^6.12.0"
yt-dlp = "^2025.2.19"
pyqt6 = "^6.8.1"
ruamel-yaml = "^0.18.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pillow = "^11.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
youtubemaster = "youtubemaster.main:main"
